Refact: rename namespace and files
