--- ScadaWeb/ScadaScheme/PlgSchSVGComp/AppCode/SchSvgComp/PlgUtils.cs
@@ -1,15 +1,0 @@
-﻿using System.Reflection;
-
-namespace Scada.Web.Plugins.SchSvgComp
-{
-	/// <summary>
-	/// Controls class loading when cloning components
-	/// </summary>
-	internal static class PlgUtils
-	{
-			
-			public static readonly SerializationBinder SerializationBinder =
-				new SerializationBinder(Assembly.GetExecutingAssembly());
-	}
-	
-}
