+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/AppCode/SchShapeComp/BoundProperties.cs
