+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/AppCode/SchShapeComp/PropertyGrid/SVGEditor.cs
@@ -1,16 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing.Design;
-using System.Linq;
-using System.Web;
 using System.Windows.Forms;
 using System.Windows.Forms.Design;
 
-namespace Scada.Web.Plugins.SchSvgComp.PropertyGrid
+namespace Scada.Web.Plugins.SchShapeComp.PropertyGrid
 {
-
-
 	public class SVGEditor : UITypeEditor
 	{
 		public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
@@ -18,7 +13,6 @@ namespace Scada.Web.Plugins.SchSvgComp.PropertyGrid
 			return UITypeEditorEditStyle.Modal;
 		}
 
-		
 		public override object EditValue(ITypeDescriptorContext context, IServiceProvider provider, object value)
 		{
 			IWindowsFormsEditorService editorService = (IWindowsFormsEditorService)provider.GetService(typeof(IWindowsFormsEditorService));
@@ -30,14 +24,14 @@ namespace Scada.Web.Plugins.SchSvgComp.PropertyGrid
 				FrmCustomShape frmCustomShape;
 				if (currentSvg is string)
 				{
-					frmCustomShape =   new FrmCustomShape(currentSvg);
+					frmCustomShape = new FrmCustomShape(currentSvg);
 				}
 				else
 				{
 					frmCustomShape = new FrmCustomShape();
 				}
 
-					
+
 				DialogResult dialogResult = frmCustomShape.ShowDialog();
 
 				if (dialogResult == DialogResult.OK)
