+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/PlgSchShapeComp.csproj
@@ -12,7 +12,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Scada.Web</RootNamespace>
-    <AssemblyName>PlgSchSVGComp</AssemblyName>
+    <AssemblyName>PlgSchShapeComp</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <UseIISExpress>true</UseIISExpress>
     <Use64BitIISExpress />
@@ -77,35 +77,35 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Content Include="plugins\SchSvgComp\css\svgcomp.css" />
-    <Content Include="plugins\SchSvgComp\js\svgcomprender.js" />
+    <Content Include="plugins\SchShapeComp\css\shapecomp.css" />
+    <Content Include="plugins\SchShapeComp\js\shapecomprender.js" />
     <None Include="Resources\svg.png" />
     <None Include="Resources\polygon.png" />
     <Content Include="Web.config" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AppCode\PlgSchSVGCompSpec.cs" />
-    <Compile Include="AppCode\SchSvgComp\BoundProperties.cs" />
-    <Compile Include="AppCode\SchSvgComp\CustomSVG.cs" />
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\Conditions\ColorCondition.cs" />
-    <Compile Include="AppCode\SchSvgComp\Polygon.cs" />
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\Conditions\PolygonCondition.cs" />
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\Forms\FrmCustomShape.cs">
+    <Compile Include="AppCode\PlgSchShapeCompSpec.cs" />
+    <Compile Include="AppCode\SchShapeComp\BoundProperties.cs" />
+    <Compile Include="AppCode\SchShapeComp\CustomSVG.cs" />
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\Conditions\ColorCondition.cs" />
+    <Compile Include="AppCode\SchShapeComp\Polygon.cs" />
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\Conditions\PolygonCondition.cs" />
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\Forms\FrmCustomShape.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\Forms\FrmCustomShape.Designer.cs">
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\Forms\FrmCustomShape.Designer.cs">
       <DependentUpon>FrmCustomShape.cs</DependentUpon>
     </Compile>
-    <Compile Include="AppCode\SchSvgComp\PlgUtils.cs" />
-    <Compile Include="AppCode\SchSvgComp\PopupSize.cs" />
-    <Compile Include="AppCode\SchSvgComp\PopupSizeConverter.cs" />
-    <Compile Include="AppCode\SchSvgComp\PopupWidth.cs" />
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\NumberSelectEditor.cs" />
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\SvgShapeSelectEditor.cs" />
-    <Compile Include="AppCode\SchSvgComp\ShapeCompFactory.cs" />
-    <Compile Include="AppCode\SchSvgComp\ShapeCompLibSpec.cs" />
-    <Compile Include="AppCode\SchSvgComp\PropertyGrid\SVGEditor.cs" />
-    <Compile Include="AppCode\SchSvgComp\SvgShape.cs" />
+    <Compile Include="AppCode\SchShapeComp\PlgUtils.cs" />
+    <Compile Include="AppCode\SchShapeComp\PopupSize.cs" />
+    <Compile Include="AppCode\SchShapeComp\PopupSizeConverter.cs" />
+    <Compile Include="AppCode\SchShapeComp\PopupWidth.cs" />
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\NumberSelectEditor.cs" />
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\SvgShapeSelectEditor.cs" />
+    <Compile Include="AppCode\SchShapeComp\ShapeCompFactory.cs" />
+    <Compile Include="AppCode\SchShapeComp\ShapeCompLibSpec.cs" />
+    <Compile Include="AppCode\SchShapeComp\PropertyGrid\SVGEditor.cs" />
+    <Compile Include="AppCode\SchShapeComp\SvgShape.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -124,7 +124,7 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="plugins\SchSvgComp\lang\" />
+    <Folder Include="plugins\SchShapeComp\lang\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ScadaSchemeCommon\ScadaSchemeCommon.csproj">
