+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/Properties/Resources.resx
