+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/Resources/polygon.png
