+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/Resources/svg.png
