+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/Web.Debug.config
