+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/Web.Release.config
