+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/Web.config
