+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/compilerconfig.json
