+++ ScadaWeb/ScadaScheme/PlgSchShapeComp/packages.config
